// lib/api/model/responseRecipeDto.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * Pina API
 * OpenAPI spec version: 1.0
 */
import type { IngredientResponseDto } from "./ingredientResponseDto";

export interface ResponseRecipeDto {
  recipeId: string;
  familyId: string;
  name: string;
  description?: string;
  tags: string[];
  servings: number;
  timeMinutes: number;
  rating: number;
  ingredients: IngredientResponseDto[];
  steps: string[];
  createdAt?: string;
  updatedAt?: string;
}